// source/AGSWebClient/packages/evidenceCore/queries/types.ts
/* 
  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  
  Licensed under the Apache License, Version 2.0 (the "License").
  You may not use this file except in compliance with the License.
  You may obtain a copy of the License at
  
      http://www.apache.org/licenses/LICENSE-2.0
  
  Unless required by applicable law or agreed to in writing, software
  distributed under the License is distributed on an "AS IS" BASIS,
  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  See the License for the specific language governing permissions and
  limitations under the License.
*/

export enum QueryType {
    SEARCH_EVIDENCE = 'SearchEvidence',
    GET_EVIDENCE = 'GetEvidence',
    GET_SCHEMA_DETAILS = 'GetSchemaDetails',
    GET_EVIDENCE_VERIFICATION_STATUS = 'GetEvidenceVerificationStatus',
    LIST_EVIDENCE_PROVIDERS = 'ListEvidenceProviders',
    GET_EVIDENCE_PROVIDER = 'GetEvidenceProvider',
    LIST_REVISIONS = 'ListRevisions',
    GET_REVISION = 'GetRevision',
    GET_REVISION_VERIFICATION_STATUS = 'GetRevisionVerificationStatus',
}

export enum MutationType {
    CREATE_EVIDENCE = 'CreateEvidence',
    CREATE_EVIDENCEPROVIDER = 'CreateEvidenceProvider',
    GENERATE_ATTACHMENT_LINK = 'GenerateAttachmentLink',
    CREATE_SCHEMA = 'CreateSchema',
}